Move setup keys to Settings
